fix payment feature

+++ src/main/java/com/javaintensive/telegrambot/modelMock/paymentserviceMock/payment/PaymentApiResponse.java
@@ -3,11 +3,15 @@ package com.javaintensive.telegrambot.modelMock.paymentserviceMock.payment;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class PaymentApiResponse {
     private PaymentStatus status;
     private Long orderId;
     private String date;
-    private Integer price;
+    private BigDecimal price;
+
+
 }
